+++ ssm/ssmShapeModelRegistrationMethod.hxx
