+++ ssm/ssmShapeModelToImageMetric.h
