+++ ssm/ssmShapeModelToImageMetric.hxx
